// src/navigation/types.ts
export type RootStackParamList = {
  Tabs: undefined;
  GoalDetail: { id?: string; title?: string } | undefined;
  TodayDetails: undefined;
  ScheduleWeek: undefined;
};

export type TabsParamList = {
  Home: undefined;
  Calendar: undefined;
  Goals: undefined;
  Ranking: undefined;
  Profile: undefined;
};